Fix macro name tokenization issue that could result in skipping some MML.
context: https://github.com/atsushieno/mugene/issues/11#issuecomment-113815000

When there is any partially-matching macro name, it incorrectly consumed
the text buffer to the partially-matching location e.g. when "cde" is being
resolved and there is variable "dem_accent", then "de" is a partial match.
In the end, 'd' was resolved as expected, but the next 'e' was already
consumed and was skipped.