Rewrite MmlResolvedEvent sorting and fixed some issues.
The previous sorter was not working as expected to sort events in the
specified order. One of the example was this:

1	CH1 V127 E127
	@48 o5 c2 d4 e4
	@1 o6 c0a2c0g0b2  c0g0>c1

It somehow put "c0" in front of "@1".

Why? The implementation-dependent sorter.

It was sorting based on the event ticks. It collected the events, collecting
ticks for each event, then ordered int ticks, then recollected events based
on the sorted int ticks.

But it was using Sort(), and for the same tick events, the result orders
were not guaranteed. Hence, in the example above, "@1" somehow came first
before "c0". It is implementation dependent, and maybe didn't happen on
Mono before.
Since mono has been replacing a lot of BCL code with Microsoft
referencesource, some internal changes would have happened and it uncovered
the issue. But that's just a guess. The previous code was implementation
dependent and was buggy anyways.

Now it preserves the events in the specified order, so the bug won't happen.